task6.c

+++ Lab03/task6.c
@@ -16,7 +16,7 @@ int main(){
     int fileOut=open("foorev2", O_CREAT | O_RDWR, 0760 );//creates a file foorev2 with rwxrw---- permissions
     char buffer[1];//one byte buffer
     int rbyte;//number of bytes read
-    int i=0;//initializes the file size
+    off_t i=0;//initializes the file size
     
     //determines the size of the file (to set the offset correctly)
     while((read(fileIn,buffer,1))>0)
